refactor: convert Searchbar to Typescript (#1037)
