--- packages/strapi-design-system/src/Searchbar/SearchForm.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const SearchForm = (props) => <form {...props} role="search" />;
