+++ packages/strapi-design-system/src/Searchbar/SearchForm.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export type SearchFormProps = Omit<React.FormHTMLAttributes<HTMLFormElement>, 'role'>;
+
+export const SearchForm = (props: SearchFormProps): JSX.Element => <form {...props} role="search" />;
