+++ packages/strapi-design-system/src/Searchbar/index.ts
